moved Notifications class into notifications.py
