+++ arm/models/models.py
@@ -19,47 +19,6 @@ import arm.config.config as cfg
 from arm.models.job import Job
 
 
-class Notifications(db.Model):
-    """
-    Class to hold the A.R.M notifications
-    """
-    id = db.Column(db.Integer, autoincrement=True, primary_key=True)
-    seen = db.Column(db.Boolean)
-    trigger_time = db.Column(db.DateTime)
-    dismiss_time = db.Column(db.DateTime)
-    title = db.Column(db.String(256))
-    message = db.Column(db.String(256))
-    diff_time = None
-    cleared = db.Column(db.Boolean, default=False, nullable=False)
-    cleared_time = db.Column(db.DateTime)
-
-    def __init__(self, title=None, message=None):
-        self.seen = False
-        self.trigger_time = datetime.datetime.now()
-        self.title = title
-        self.message = message
-
-    def __repr__(self):
-        return f'<Notification {self.id}>'
-
-    def __str__(self):
-        """Returns a string of the object"""
-
-        return_string = self.__class__.__name__ + ": "
-        for attr, value in self.__dict__.items():
-            return_string = return_string + "(" + str(attr) + "=" + str(value) + ") "
-
-        return return_string
-
-    def get_d(self):
-        """ Returns a dict of the object"""
-        return_dict = {}
-        for key, value in self.__dict__.items():
-            if '_sa_instance_state' not in key:
-                return_dict[str(key)] = str(value)
-        return return_dict
-
-
 class SystemInfo(db.Model):
     """
     Class to hold the system (server) information
